+++ jni/java/app/opendocument/core/PageLayout.java
@@ -6,11 +6,18 @@ public final class PageLayout {
   public final Measure height;
   public final PrintOrientation printOrientation;
   public final DirectionalMeasure margin;
+  public final Color backgroundColor;
 
-  PageLayout(Measure width, Measure height, int printOrientation, DirectionalMeasure margin) {
+  PageLayout(
+      Measure width,
+      Measure height,
+      int printOrientation,
+      DirectionalMeasure margin,
+      Color backgroundColor) {
     this.width = width;
     this.height = height;
     this.printOrientation = PrintOrientation.fromNative(printOrientation);
     this.margin = margin;
+    this.backgroundColor = backgroundColor;
   }
 }